Added Oxide Updates
This adds oxide update checking. This was a needed requirement due to
Oxide needing to be updated any time the server is. If the Oxide is
enabled, the script will only check for Oxide updates and update when
Oxide is ready instead of the game update. This will prevent the game
from updating and an old version of Oxide overwriting.  If Oxide is
disabled, the normal update process is used.