Oxide Current Version Check Updates
This updates how the latest version of Oxide is checked. It will look
for log files until it finds the latest Oxide available file, and then
will look for the latest version of Oxide loaded from that log file.  If
the log file is not found it will be written with either the saved
version from the last time the update check was ran or 0.0.0 if the
update check is running for the first time.  The version is remembered
so it can be properly written and saved if the log rotates while the
script is still running.